requirements.txt
# Core dependencies
opencv-python==4.8.1.78
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Screen capture
mss>=9.0.0

# AI and LLM
openai>=1.0.0
python-dotenv>=1.0.0

# Window detection (platform-specific)
pywin32>=306; sys_platform == "win32"  # Windows

# Poker hand evaluation
treys>=0.1.5

# Tournament and dynamic detection
dataclasses>=0.6; python_version<"3.7"
enum34>=1.1.10; python_version<"3.4"

# Utilities
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development and Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.0.0
mypy>=1.0.0
bandit>=1.7.0
